test: Merge test modules for default domains and FractureNetworks
